Cygwin: configure: add possibility to skip build of cygserver and utils
This patch adds configure options allowing to disable build of cygserver
and Cygwin utilities. This is useful when one needs to build only
cygwin1.dll and crt0.o with stage1 compiler that is not yet capable of
linking executables as it is missing cygwin1.dll and crt0.o.

Signed-off-by: Radek Bartoň <radek.barton@microsoft.com>
(cherry picked from commit b6e108502ccacac687852a01dd5c321710b69f17)